build_sdk: add SMP configs not SMP boards
This means that the board names are the same, but we
build extra configurations of the boards with different
kernel configurations for SMP support. This is cleaner, as
we were finding that renaming the boards was broken a lot
of our build systems code which needed to be duplicate for
new boards.

I'm not entirely convinced this is the best possible solution,
but it is the nicest we can think of.

Signed-off-by: Julia Vassiliki <julia.vassiliki@unsw.edu.au>